fix(hydro_lang): avoid crashes and miscompilations with unused locations (#1641)
Previously, if a location was not used in the graph, we would crash or
use faulty IDs in the generated code. This also changes semantics
slightly so that if there is a cluster with no assigned computation, it
will be treated as an empty cluster at runtime (the list of cluster IDs
will be an empty slice).