fix(hydro_lang): correctly resolve `__staged` paths for the crate being launched
Previously, if launching code from `hydro_test`, we would fail to resolve relative paths originating inside `hydro_test` code, since the `__staged` module in the trybuild is empty (unless in test mode, which is why we missed this). Now, when we are not in test mode, we set up the trybuild `__staged` to redirect to the `__staged` of the crate the trybuild is being generated for, so that the correct path is resolved.