refactor(hydro_test): migrate tests from hydro_test_local
We are deprecating the macro entrypoint API, since it is brittle and not compatible with distributed deployments. Move the useful tests into hydro_test.

Because the macro no longer pulls in `hydro_lang`, this also reduces the features pulled in by workspace unification, which causes compilation bugs when running tests for `hydro_lang`. So this PR also includes some fixes to reduce brittleness there.

Towards #1872