feat(hydro_lang): upgrade Stageleft to eliminate `__staged` compilation during development
Before Stageleft 0.9, we always compiled the `__staged` module in stage 0, which resulted in significant compilation penalties and Rust Analyzer thrashing since any file changes triggered a re-run of the `build.rs`. With Stageleft 0.9, we can defer compiling this module to the trybuild stage 1.

Stageleft 0.9 also cleans up how paths are rewritten to use the `__staged` module, so we can simplify our logic as well. The only significant rewrite remaining is when running unit tests, where we have to regenerate `__staged` to access test-only module, and therefore have to rewrite all paths to use that module.

Finally, in the spirit of improving compilation efficiency, we disable incremental builds for trybuild stage 1. We generate files with hash based on contents, so we were never benefitting from incremental compilation anyways. This reduces the disk space used significantly.

Towards #1872