feat(hydro_lang)!: add stream markers for tracking non-deterministic retries
This introduces an additional type paramter to `Stream` called `Retries`, which tracks the presence (or lack) of non-determinstic retries in the stream. `ExactlyOnce` means that each element has deterministic order, while `AtLeastOnce` means that there may be non-deterministic duplicates.

A `TotalOrder, AtLeastOnce` stream describes elements with consecutive duplication, but deterministic order if we ignore those immediate elements. A `NoOrder, AtLeastOnce` stream has set semantics.

Also fixes a bug in the return type for `*_keyed_*`, where the output type was previously `TotalOrder` but now is `NoOrder`. We stream the results of a keyed aggregation out of a `HashMap`, so the order will indeed be non-deterministic.

Fixes #1863