refactor(hydro_lang)!: invert external sources and clean up locations in IR (#1958)
First, instead of creating external sources by invoking
`external.source_bincode_external(&p)`, we switch the API to
`p.source_bincode_external(&external)` for symmetry with `source_iter`
and `source_stream`.

The other, much larger change is to clean up how the IR handles external
inputs and outputs and keeps track of locations. First, we introduce
`HydroNode::ExternalInput` and `HydroLeaf::SendExternal` as specialized
nodes for these, so that we no longer create dummy sources / sinks.

Then, we eliminate places where we have multiple sources of truth for
where the output of an IR node is located, by instead referring to the
metadata. Because it is easy in optimizer rewrites to corrupt this
metadata, we also add a flag to `transform_bottom_up` that lets
developers enable a metadata validity check. We disable it in most
transformations for performance, but enable it in the decoupling
rewrites since it manipulates locations in complex ways.