refactor(hydro_lang): separate externals from other location kinds, clean up network operators (#1962)
First, we remove externals from `LocationId`, to ensure that a
`LocationId` only is used for locations where we can concretely place
compiled logic. This simplifes a lot of pattern matching where we wanted
to disallow externals.

Keys on network inputs / outputs (`from_key` and `to_key`) are only
relevant to external networking. We extract the logic to instantiate
external network collections, so that the core logic does not need to
deal with keys.