feat(hydro_lang): add `Stream::scan_keyed` and `Stream::fold_keyed_early_stop`
In preparation for a keyed variant of `Stream` that allows richer reasoning about interleaved ordering, we add a keyed variant of scan and a keyed variant of fold that are suitable for unbounded streams.

The latter is particularly interesting, as it allows an early-termination (inspired by Free Termination) to yield folded results (since without early termination, we can never yield anything if the input is unbounded). With this operator, the entire HTTP example is key-able which will eventually let us get rid of the `unsafe` blocks surrounding it.