refactor(hydro_lang)!: simplify networking APIs according to Process/Cluster types (#1975)
Breaking change: when sending to a cluster, you must use
`demux_bincode`. The `*_anonymous` APIs have been removed in favor of
the `.values()` API on keyed streams. This also eliminates the
`send_bytes` APIs, in favor of the bidirectional external client APIs.

This refactors the networking APIs to rely less on complex traits and
instead use `Process` and `Cluster` types to determine the input /
output type restrictions. We also now emit `KeyedStream` whenever the
sender is a `Cluster`.