refactor(hydro_lang)!: nondet instead of unsafe, snapshot instead of latest_tick, batch instead of tick_batch (#1990)
Pulled together into one big PR to avoid conflicts, this makes three
significant changes to the core Hydro APIs. No semantic changes, only
syntax and naming.

1. Instead of non-deterministic functions being marked with the `unsafe`
keyword, which conflates non-determinism with memory unsafety, we now
pass around "non-determinism guards" that act as tokens forcing
developers to be aware of the non-determinism. With the VSCode Highlight
extension, we preserve highlighting of these instances.
2. We unify naming for "collection-like" types (Stream, KeyedStream) to
use `.batch` to batch elements
3. We unify naming for "floating-value-like" types (Singleton, Optional,
KeyedSingleton, KeyedOptional) to use `.snapshot` to grab an instance of
the value (or its contents)