refactor(hydro_lang)!: remove `KeyedOptional` (#2011)
Keyed optionals don't make sense to distinguish from keyed singleton,
since a null optional is equivalent to the key not being present at all.
So we can capture all cases with just a keyed singleton.

Eventually... we may want to re-distinguish these if we have keys with a
null value. But right now we don't need that so avoid unnecessary code.