feat(hydro_lang)!: refine types when values of a `KeyedSingleton` are immutable
There are many cases where once a value is released into a `KeyedSingleton`, it will never be changes. In such cases, we can permit APIs like `.entries` even if the set of entries is growing.

We use this type to improve the API surface for looking up values for a request. Now, a `KeyedSingleton` of requests can look up values from another `KeyedSingleton`.