refactor(hydro_lang): cleanly separate generic deployment from `hydro_deploy` backend
Now, the `builder` module captures all the logic and trait definitions for deploying Hydro program to a generic backend. All the code for deploying _via_ Hydro Deploy lives in `deploy`.

Also moves the `deploy_runtime` module to be inside `deploy` as a private module, to clean up the public API.