fix(hydro_lang)!: fix cardinality for `Optional::or`
Using `HydroNode::Chain` is very dangerous for singletons / optionals, because it can lead to cardinality > 1 within a single batch, which breaks a fundamental invariant.

This introduces a new `HydroNode::ChainFirst` operator that only emits the first value from the chain. This is paired with flag to the DFIR operator for this behavior.

We also rewrite some `Singleton` logic to use `Optional` under the hood, which reduces the places where we deal with chaining in the IR.