refactor(hydro_lang)!: migrate Hydro IR to have Flo semantics at all levels
Previously, the Hydro -> Hydro IR layer did a bunch of work to translate between Flo semantics and DFIR semantics, so that Hydro IR -> DFIR was generally 1:1. In preparation for the simulator and DFIR's migration to Flo semantics, we now use Flo semantics in the Hydro IR (top level operators do not reset their state on batch boundaries), and shift the Flo -> DFIR semantics translation to the IR -> DFIR layer.

This also comes with a breaking API change to clean up naming and avoid function overloading. In particular, `batch` / `snapshot` are renamed to `batch_atomic` and `snapshot_atomic` for the case where the batched result is supposed to be in-sync with the atomic execution context.

There were a couple of bugs during the transition to the new IR semantics that were not caught by existing unit tests, so I added additional tests that cover each of them.

Also makes some minor improvements the way:
- Groups together definitions for `resolve_futures` and `resolve_futures_ordered`
- Removes various unnecessary `NoAtomic` trait bounds