refactor(hydro_lang)!: shift `Stream::max_by_key` to `KeyedSingleton::get_max_key`, which is actually deterministic
The previous API would be non-deterministic if there are two elements in the stream that have the same computed key but different values (since we would randomly pick one of them). So we shift the API to `KeyedSingleton`, where we are guaranteed to have only a single element per key.