refactor(hydro_lang): don't pass around locations when emitting DFIR
Because the IR now includes the location of the output, we can just use that in the codegen process. This cleans up the code and also prepares us for simulator codegen, where the output DFIR will depend on the tick.

Also fixes a bug I found in top-level `unique` while refactoring, and adds a corresponding unit test.