feat(hydro_lang): replay failed simulation instances and show logs
Previously, a failed simulation would only show the backtrace, and would have to be re-run with `HYDRO_SIM_LOG=1` to display the simulation steps. This would result in lots of log pollution since logging would also be enabled for passing instances.

Now, we've made some tweaks to bolero so that it re-executes the test after finding a failing input, passing an `is_replay` flag. We use this to enable rich logging, so that the failure is displayed with all the simulation steps that led to it.