test(hydro_lang): check that simulated unordered batching generates out-of-order cases
Until we can simulate `assume_ordering` inside a tick, we need some way for getting the decisions out. All the `*_commutative` APIs currently call `assume_ordering`, but we introduce an `_trusted` variant for `min` / `max` that bypasses `assume_ordering` because we have carefully checked that the function is indeed commutative and we do not need to simulate shuffles.

Eventually, the decision whether we need to simulate shuffled executions of `_commutative` methods will be driven by algebraic verification, so that user code can similarly benefit from the reduction in search space.