feat(hydro_lang): share `__staged` across trybuild targets
Previously, each trybuild target (in test-mode) would have its own copy of the `__staged` module containing the entire sources of the original crate. This creates a huge compilation burden that is repeated on every deployment.

With an accompanying change in Stageleft, we are now able to share the `__staged` module in the common `lib.rs`. This also moves the simulation dylibs to be compiled as an example so that they do not conflict with the caching of the shared `lib.rs`.