feat(hydro_lang): support singletons in simulator
The simulator uses a different representation of singletons than production, streaming every intermediate state instead of just the final value for the current DFIR batch. This lets us simulate all possible snapshots when that happens.

To support `Location::singleton`, which (lazily) spins in production, we instead introduce a proper `SourceSingleton` IR node which only (appropriately) only releases a single item in the simulator.

To simulate snapshots, we have to decide which version of the singleton to advance to. This can include "stutters", where the same version is released multiple times. To avoid an infinite search space, stutters are only allowed when some other input to the tick has made a nontrivial decision (one that makes progress consuming its input). Therefore, if none of the other inputs make progress, the singleton has to make progress.