feat(hydro_lang): support `Stream::count` in simulator
This also tweaks the behavior of `reduce_commutative_idempotent_trusted` to make sure that we simulate different orders of intermediate states when the input is unbounded / asynchronous. Right now, we do not actually have any tests for this, since we do not support shuffling unbounded streams, but this will be supported soon.