test: include Hydro template sources into workspace tree and test example
Since none of the template sources rely on the template variables, we can include them as-is into the workspace tree so that we get code completion and can run tests without having to manually generate the template.

Also fixes a bug with Hydro not supporting custom `lib.path` (with accompanying Stageleft fix).