feat(hydro_lang): add `KeyedStream` APIs for compound keys and sending with keys (#2298)
Previously, the only way to send a `KeyedStream` _including_ the keys
was to get the entries (which would erase the order).

Now, we add APIs to prefix/unprefix the key (prepend a tuple element),
and then add a `KeyedStream::send_bincode` API that sends the entire set
of entries with the sender ID prefixed in the return type. We also add
an inverse `KeyedStream::demux_bincode` where the input keys are a
compound of `(MemberId, Key)` and the recipient gets a
`KeyedStream<Key>`.