refactor(hydro_std): port `bench_client` to use sliced!, introduce `Stream::merge_ordered`
This also fixes a bug where the throughput reset timer would cause a batch of payloads to be not counted in the throughput.

We add support for tick-cycles / `use::state` to `Optional` and `KeyedSingleton` to support this code more cleanly.

To further simplify the code, we add `merge_ordered` to combine two streams while preserving order. This makes it possible to implement throughput windowing without slices.